# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type AiChatMessageRole =
  (typeof AiChatMessageRole)[keyof typeof AiChatMessageRole];

export const AiChatMessageRole = {
  user: "user",
  assistant: "assistant",
  system: "system",
} as const;

export interface AiChatMessage {
  role: AiChatMessageRole;
  content: string;
}

export interface AiChatRequest {
  /** @minItems 1 */
  messages: AiChatMessage[];
  stream?: boolean;
}

export interface TokenUsage {
  promptTokens: number;
  completionTokens: number;
  totalTokens: number;
}

export interface AiChatResponse {
  message: AiChatMessage;
  usage?: TokenUsage;
  model: string;
}

export type ContentType = (typeof ContentType)[keyof typeof ContentType];

export const ContentType = {
  blog_post: "blog_post",
  social_media: "social_media",
  email_campaign: "email_campaign",
  landing_page_copy: "landing_page_copy",
  ad_copy: "ad_copy",
  case_study: "case_study",
} as const;

export type ContentGenerationRequestTone =
  (typeof ContentGenerationRequestTone)[keyof typeof ContentGenerationRequestTone];

export const ContentGenerationRequestTone = {
  professional: "professional",
  conversational: "conversational",
  authoritative: "authoritative",
  inspiring: "inspiring",
} as const;

export interface ContentGenerationRequest {
  type: ContentType;
  topic: string;
  tone?: ContentGenerationRequestTone;
  keywords?: string[];
  targetAudience?: string;
  additionalContext?: string;
}

export interface ContentGenerationResponse {
  content: string;
  title?: string;
  type: ContentType;
  model: string;
  usage?: TokenUsage;
}

export type BusinessDomain =
  (typeof BusinessDomain)[keyof typeof BusinessDomain];

export const BusinessDomain = {
  automation: "automation",
  marketing: "marketing",
  operations: "operations",
  growth_strategy: "growth_strategy",
  ai_integration: "ai_integration",
  brand_positioning: "brand_positioning",
} as const;

export interface BusinessAssistantRequest {
  question: string;
  domain?: BusinessDomain;
  businessContext?: string;
  industry?: string;
}

export interface BusinessAssistantResponse {
  answer: string;
  recommendations?: string[];
  nextSteps?: string[];
  domain?: BusinessDomain;
  model: string;
  usage?: TokenUsage;
}