# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: AI chat completion
      description: Send messages to the STAGEONE AI assistant and receive a response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatRequest"
      responses:
        "200":
          description: Successful AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: AI service error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AiChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
      required:
        - role
        - content
    AiChatRequest:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AiChatMessage"
          minItems: 1
        stream:
          type: boolean
          default: false
      required:
        - messages
    AiChatResponse:
      type: object
      properties:
        message:
          $ref: "#/components/schemas/AiChatMessage"
        usage:
          $ref: "#/components/schemas/TokenUsage"
        model:
          type: string
      required:
        - message
        - model
...
    TokenUsage:
      type: object
      properties:
        promptTokens:
          type: integer
        completionTokens:
          type: integer
        totalTokens:
          type: integer
      required:
        - promptTokens
        - completionTokens
        - totalTokens