# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Send messages to the STAGEONE AI assistant and receive a response
 * @summary AI chat completion
 */

export const aiChatBodyStreamDefault = false;

export const AiChatBody = zod.object({
  messages: zod
    .array(
      zod.object({
        role: zod.enum(["user", "assistant", "system"]),
        content: zod.string(),
      }),
    )
    .min(1),
  stream: zod.boolean().default(aiChatBodyStreamDefault),
});

export const AiChatResponse = zod.object({
  message: zod.object({
    role: zod.enum(["user", "assistant", "system"]),
    content: zod.string(),
  }),
  usage: zod
    .object({
      promptTokens: zod.number(),
      completionTokens: zod.number(),
      totalTokens: zod.number(),
    })
    .optional(),
  model: zod.string(),
});

/**
 * Generate high-quality content for various business use cases
 * @summary Generate marketing or business content
 */
export const generateContentBodyToneDefault = `professional`;

export const GenerateContentBody = zod.object({
  type: zod.enum([
    "blog_post",
    "social_media",
    "email_campaign",
    "landing_page_copy",
    "ad_copy",
    "case_study",
  ]),
  topic: zod.string(),
  tone: zod
    .enum(["professional", "conversational", "authoritative", "inspiring"])
    .default(generateContentBodyToneDefault),
  keywords: zod.array(zod.string()).optional(),
  targetAudience: zod.string().optional(),
  additionalContext: zod.string().optional(),
});

export const GenerateContentResponse = zod.object({
  content: zod.string(),
  title: zod.string().optional(),
  type: zod.enum([
    "blog_post",
    "social_media",
    "email_campaign",
    "landing_page_copy",
    "ad_copy",
    "case_study",
  ]),
  model: zod.string(),
  usage: zod
    .object({
      promptTokens: zod.number(),
      completionTokens: zod.number(),
      totalTokens: zod.number(),
    })
    .optional(),
});

/**
 * Get AI-powered business strategy and automation recommendations
 * @summary Business strategy assistant
 */
export const BusinessAssistantBody = zod.object({
  question: zod.string(),
  domain: zod
    .enum([
      "automation",
      "marketing",
      "operations",
      "growth_strategy",
      "ai_integration",
      "brand_positioning",
    ])
    .optional(),
  businessContext: zod.string().optional(),
  industry: zod.string().optional(),
});

export const BusinessAssistantResponse = zod.object({
  answer: zod.string(),
  recommendations: zod.array(zod.string()).optional(),
  nextSteps: zod.array(zod.string()).optional(),
  domain: zod
    .enum([
      "automation",
      "marketing",
      "operations",
      "growth_strategy",
      "ai_integration",
      "brand_positioning",
    ])
    .optional(),
  model: zod.string(),
  usage: zod
    .object({
      promptTokens: zod.number(),
      completionTokens: zod.number(),
      totalTokens: zod.number(),
    })
    .optional(),
});